requirements.txt
kafka-python
tensorflow
matplotlib
six
numpy